Fold PocoEdgeCaseTests into PocoReadTests / InsertBinaryPocoTests
Two fixtures lived in PocoEdgeCaseTests.cs:

  - PocoEdgeCaseTests          — unit-style POCO-shape registration smokes
  - PocoReadEdgeCaseTests      — integration MapTo / QueryAsync edge cases

Their content fits naturally in the two existing per-side integration
fixtures. Fold them so each kind of test lives next to its peers:

  - PocoReadTests gains the read-side RegisterPocoType<T> throws tests
    (positional records, all-init-only, all-private-setter, all-readonly,
    abstract, fields-only, static-only, write-only) and every MapTo /
    QueryAsync edge case (NotMappedProperty, indexer, init/readonly/
    private-setter skip, widening throw, object property, derived,
    multiple types per reader, before-Read, after-end-of-stream,
    required members, decimal-without-custom-decimals, unregistered-on-
    first-yield, cancellation mid-iteration, server error mid-stream,
    failed-insert-rollback proven via QueryAsync).
  - InsertBinaryPocoTests gains the RegisterBinaryInsertType<T>
    registration smokes for shapes that are insert-mappable but not
    read-mappable (positional records, init-only, private-setter,
    read-only, abstract).

Five tests dropped as redundant with the integration coverage that's
now in the same file: RecordWithMixedAccessors / PrivateSetter /
MixedReadOnlyAndReadWrite registration smokes (all covered by the
runtime MapTo skip test), DerivedType registration smoke (covered by
the MapTo derived test), and RequiredMembers registration smoke
(covered by the MapTo required-members test).

PocoEdgeCaseTests.cs is removed.
